Redo SiegeManager
